backend/requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
black==23.7.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
FastAPI-SQLAlchemy==0.2.1
flake8==6.1.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
psycopg2==2.9.7
pycodestyle==2.11.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyflakes==3.1.0
pytest==7.4.1
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.23.2
